added bag id popup
