+++ 3DAmsterdam/Assets/T3D/Scripts/UI/BagIDHoverButton.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BagIDHoverButton : InfoHoverButton
+{
+    private void Awake()
+    {
+        defaultPopupText = "Bag ID: " + T3DInit.Instance.BagId;
+    }
+}
