+++ 3DAmsterdam/Assets/T3D/Scripts/UI/BagIDHoverButton.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c36fc96fdcbd1467fb7ccd55d43155ec
+guid: f676c44bf6e8d4c03804c37a1122dabc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
