+++ 3DAmsterdam/Assets/T3D/Scripts/UI/RestrictionPopupText.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c36fc96fdcbd1467fb7ccd55d43155ec
+guid: a6b94b1e3e3894256af921a6e925dc50
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
